[357386] Fix the usage of OBJ_DESC_PACKAGE image
The problem was that every time the PackageLabelProvider.getImage() is
called a new OBJ_DESC_PACKAGE image is created. This is not efficient
usage of resources. Moreover, the created images are not disposed later
and this could causes leaking of image handlers on OS level. 

The fix is to use the OBJ_DESC_PACKAGE image from the PDE image
repository.