[357386] Fix the usage of DESC_EXTENSIONS_OBJ image
The problem was that every time the ServiceLabelProvider.getImage()
method is called a new DESC_EXTENSIONS_OBJ image is created. This is not
efficient usage of resources. Moreover, the created images are not
disposed later and this could causes leaking of image handlers on OS
level. 

The fix is to use the DESC_EXTENSIONS_OBJ image from the PDE shared
label provider.