[357386] Remove the tight coupling to the Virgo Manifest editor
The "open bundle manifest" action is re-implemented in more generic way.
It checks for available IContentType and finds the editor that
corresponds to it. The action first try to find if there any special
content types registered for MANIFEST.MF (like the Virgo Manifest
editor). If none are found, then it tries to retrieve the content type
for the PDE Manifest. If PDE is not present, then it falls back to the
default plain text content type.